VideoGen/requirements.txt
streamlit
requests
yfinance
edge-tts
moviepy==1.0.3
matplotlib
Pillow
python-dotenv
pandas
